problems fixed

+++ src/handlers/product/deleteProductById/handler.ts
@@ -1,10 +1,21 @@
+import { HttpError } from '@fastify/sensible/lib/httpError'
 import { RouteHandlerMethod } from 'fastify'
 import { StatusCodes } from 'http-status-codes'
 import { DeleteProductByIdParams } from './parameter'
 import { DeleteProductByIdResult, DeleteProductByIdResponse } from './response'
 
-export const deleteProductByIdHandler: RouteHandlerMethod = async function (request, reply): Promise<DeleteProductByIdResponse> {
-    const params = request.query as DeleteProductByIdParams
+export const deleteProductByIdHandler: RouteHandlerMethod = async function (request, reply): Promise<DeleteProductByIdResponse | HttpError> {
+    const params = request.params as DeleteProductByIdParams
+
+    const product = await this.prisma.product.findUnique({
+        where: {
+            productId: params.productId,
+        },
+    })
+
+    if (!product) {
+        return this.httpErrors.notFound('Product not found.')
+    }
 
     await this.prisma.product.delete({
         where: {
